Factorization algorithms for a polynomial in a finite field (SFD, DDF, EDF):
Showing up next there are three different algorithms that factorizes a polynomial 
in different ways. The structure of the file is:

- 0. Auxiliary functions:

Let F be a field with |F| = q = p^w, w != 1, the function PolynomialToFx transforms
a polynomial f with alpha, x variables to a polynomial with coefficients in F.

To build F, it's posible for q to be not prime, so we need to apply the action
of a polynomial "pol" irreducible with degree w to express F as Zp/(pol).

In the case w=1 you'll have to enter pol=1 (it's not neccesary the action of 
a polynomial becouse F can be represented as Zp). 

The second auxiliary function calculates the G.C.D. in a finite field of two 
primitive polynomials with coefficients in F, where F,p and pol are described 
before.

- 1. SFD (Square-free descomposition)

Let F be a field with |F| = q = p^w and "pol" be an irreducible polynomial with
degree w with which we can express F as Zp/(pol), this algorithm returns a list
of pairs [(g[i], s[i])], where f is the multiplication of (g[i])^(s[i]), and 
every g[i] is monic and square-free.


- 2. DDF (Distinct Degree Factorization)

Let F be a field with |F| = q = p^w and "pol" be an irreducible polynomial with
degree w with which we can express F as Zp/(pol), this algorithm returns a list
of pairs [(g[i]),k[i])], where f is the multiplication of g[i], and every g[i]
can be factorized in irreducible polynomials of degree k[i].

- 3. EDF (Equal Degree Factorization)

Let F be a field with |F| = q = p^w and "pol" be an irreducible polynomial with
degree w with which we can express F as Zp/(pol), this algorithm returns a list 
with the irreducible factors of a polynomial f if we know that f can be 
factorized in k irreducible polynomials.