Berlekamp's algorithm in a finite field
The structure of the algorithm is based in:

- 0. Auxiliary functions: 

Let Q be a matrix. The function insertRow inserts to a vector "r" the first n-elements
of a row "i".

Let f be a polynomial and p be the cardinal of Zp with which we want to express f 
apllying this algorithm, the function matrixQ provides the Berkelamp matrix of 
that polynomial. 

Let "a" be a number. The function inverseModulus returns its inverse modulus "b".

Let M be a matrix of dimension n, the function permuteRow swap the row i with the row j.

Let M be a matrix of dimension n and p the cardinal of Zp, the function berlekampVectors
returns the Berlekamp vectors v(1), ... , v(n).


- 1. Final algorithm:

Let f be a square-free polynomial and p the cardinal of Zp, this algorithm completely 
factorizes f in factors contained in Zp[x].